+++ firetruck/state.h
@@ -22,11 +22,11 @@ typedef enum {
 } sound_anim_t;
 
 typedef struct {
-	i16 x;
-	i16 y;
+	i32 x;
+	i32 y;
 
-	i16 rx;
-	i16 ry;
+	i32 rx;
+	i32 ry;
 
 	bool btn_a;
 	bool btn_x;
@@ -36,8 +36,8 @@ typedef struct {
 	bool shoulder_l;
 	bool shoulder_r;
 
-	u16 throttle;
-	u16 brake;
+	i32 throttle;
+	i32 brake;
 
 	struct {
 		sound_anim_t anim;
@@ -46,11 +46,11 @@ typedef struct {
 } State;
 
 typedef struct {
-	i16 x;
-	i16 y;
+	i32 x;
+	i32 y;
 
-	i16 rx;
-	i16 ry;
+	i32 rx;
+	i32 ry;
 
 	bool btn_a;
 	bool btn_a_toggle;
@@ -61,9 +61,8 @@ typedef struct {
 	bool shoulder_l;
 	bool shoulder_r;
 
-	u16 throttle;
-	u16 brake;
-
+	i32 throttle;
+	i32 brake;
 } CurrentState;
 
 volatile extern State state;
